src/connector/requirements.txt
# MongoDB driver
pymongo>=4.5.0

# Google Cloud dependencies
google-cloud-pubsub>=2.18.0
google-cloud-firestore>=2.13.0

# Configuration and environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Logging and monitoring
structlog>=23.2.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-pymongo>=0.41b0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
mongomock>=4.1.2

# Type checking and linting
mypy>=1.5.1
types-PyYAML>=6.0.12.12
pymongo-stubs>=0.2.0
